+++ src/package.rs
@@ -1,5 +1,5 @@
+use crate::git::{GitReference, GitRepo};
 use crate::{Error, Result};
-use crate::git::GitRepo;
 
 use std::env;
 use std::fmt;
@@ -37,7 +37,6 @@ lazy_static! {
         });
     static ref VIM_PACKAGE_DIR: PathBuf = (*VIM_BASE_DIR).join("pack");
     static ref VIM_PLUGIN_DIR: PathBuf = (*VIM_BASE_DIR).join("plugin");
-
     static ref PAC_CONFIG_DIR: PathBuf = (*VIM_BASE_DIR).join(".pac");
     static ref PAC_CONFIG_FILE: PathBuf = (*PAC_CONFIG_DIR).join("paconfig.yaml");
 }
@@ -53,6 +52,8 @@ pub struct Package {
     pub idname: String,
     /// Remote url of the repo to git clone from
     pub remote: String,
+    /// The branch, tag, or commit to checkout
+    pub reference: Option<GitReference>,
     /// Install package under pack/<category>/
     pub category: String,
     pub opt: bool,
@@ -65,11 +66,18 @@ pub struct Package {
 }
 
 impl Package {
-    pub fn new(name: &str, remote: &str, category: &str, opt: bool) -> Package {
+    pub fn new(
+        name: &str,
+        remote: &str,
+        reference: Option<GitReference>,
+        category: &str,
+        opt: bool,
+    ) -> Package {
         Package {
             name: name.to_string(),
             idname: Self::idname_from_remote(remote),
             remote: remote.to_string(),
+            reference,
             category: category.to_string(),
             opt,
             load_command: None,
@@ -112,6 +120,18 @@ impl Package {
             .as_str()
             .map(|s| s.to_string())
             .ok_or(Error::Format)?;
+
+        let reference = match doc["ref"].as_hash() {
+            None => None,
+            Some(subdoc) => {
+                let (reftype, refval) = subdoc.iter().next().ok_or(Error::Format)?;
+                let to_str = |yaml: &Yaml| -> Result<String> {
+                    yaml.as_str().map(|s| s.to_string()).ok_or(Error::Format)
+                };
+                Some(GitReference::new(&to_str(reftype)?, &to_str(refval)?)?)
+            },
+        };
+
         let opt = doc["opt"].as_bool().ok_or(Error::Format)?;
         let category = doc["category"]
             .as_str()
@@ -135,6 +155,7 @@ impl Package {
             name,
             idname: Self::idname_from_remote(&remote),
             remote,
+            reference,
             category,
             opt,
             load_command: cmd,
@@ -150,6 +171,16 @@ impl Package {
         doc.insert(Yaml::from_str("remote"), Yaml::from_str(&self.remote));
         doc.insert(Yaml::from_str("category"), Yaml::from_str(&self.category));
         doc.insert(Yaml::from_str("opt"), Yaml::Boolean(self.opt));
+
+        if let Some(ref gitref) = self.reference {
+            let mut subdoc = Hash::new();
+            subdoc.insert(
+                Yaml::from_str(&gitref.kind.to_string()),
+                Yaml::from_str(&gitref.value),
+            );
+            doc.insert(Yaml::from_str("ref"), Yaml::Hash(subdoc));
+        }
+
         if let Some(ref c) = self.load_command {
             doc.insert(Yaml::from_str("on"), Yaml::from_str(c));
         }
@@ -170,9 +201,15 @@ impl Package {
     /// Returns absolute path to directory where plugin can be installed
     pub fn path(&self) -> PathBuf {
         if self.opt {
-            VIM_PACKAGE_DIR.join(&self.category).join("opt").join(&self.name)
+            VIM_PACKAGE_DIR
+                .join(&self.category)
+                .join("opt")
+                .join(&self.name)
         } else {
-            VIM_PACKAGE_DIR.join(&self.category).join("start").join(&self.name)
+            VIM_PACKAGE_DIR
+                .join(&self.category)
+                .join("start")
+                .join(&self.name)
         }
     }
 
@@ -204,8 +241,8 @@ impl Package {
 }
 
 impl GitRepo for Package {
-    fn path_info(&self) -> (&str, PathBuf) {
-        (&self.remote, self.path())
+    fn clone_info(&self) -> (&str, PathBuf, Option<GitReference>) {
+        (&self.remote, self.path(), self.reference.clone())
     }
 }
 
@@ -365,7 +402,6 @@ where
     })?;
     Ok(())
 }
-
 
 #[cfg(test)]
 mod tests {